pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "opsbutler"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]
